have cli watch template file for changes (#91)
There was a bug previously where if this was run from a subdirectory, it
would always use the default remote template, since it wouldn't find the
local one where it expected. That's fixed in this PR.

Also, while it would detect a change in specified template file in the
Stoat config file, it wouldn't update automatically on changes within
the template itself, which would require restarting `stoat local`. This
wouldn't be that big of a deal except the one of the biggest reasons to
use local mode is to fine tune the template.

I created https://github.com/stoat-dev/stoat-action/issues/90 to track
an issue that would make it trivial to locally iterate on templates
including non-`static_hosting` plugins. Maybe we could even have a copy
button on debug pages next to the merged configs that just gives you the
command to run locally to preview that exact config.